requirements.txt
pandas>=2.0
scikit-learn>=1.6.1
fastapi>=0.110
uvicorn>=0.29
pyyaml>=6.0
joblib>=1.3
boto3>=1.34
python-dotenv>=1.0
numpy>=1.26
pytest>=8.0
httpx>=0.27
xgboost>=2.0
lightgbm>=4.0
catboost>=1.2
scipy>=1.11
